Do not extract comment only interpolations
I have seen comments (multiline ones) which are wrapped in ERB templates
like this:

<%# here is some comment
    it carries on to this line
    and more
%>

Doing a comment this way does not include it in the final HTML which
would be the case when using html comments. If ruumba extracts this,
since it is multiline, it is actually invalid ruby code.

I have updated the parsing code to just throw away lines which begin
with a '#'.