Use pwd after switching to temporary directory to replace paths in output
If the tmp dir is a symlink to another directory, after the code in
ruumba which does Dir.chdir(tmp) runs, we are not actually in the
directory returned from `FileUtils.mkdir_p(File.dirname(n))`. Delay
creating the regex replacements until we switch directories, so we can
be sure we have the correct directory name to replace in the output.

Closes #19.