Fix parsing comments when auto-correct is enabled.
The markers we use to delimit where the pieces of code were inside the
RHTML get lost because of a logic error in the parser. The region close
marker does not be included and once this string it returned from the
parser, the replace code will be unable to put the code back in the
correct place.