Strip trailing whitespace and newlines when not autocorrecting.
Since we extract the ruby code from inside the template and run rubocop
on it, keeping the column positions (filling with whitespace) so that we
can find where the code came from when auto correcting, most layout and
style cops don't make much sense with ruumba.

If we are not auto-correcting (which is experimental currently anyway),
we can strip out the trailing spaces and the trailing newlines to avoid
those violations. We will still get indention cop violations etc, but
perhaps this at least gives a slightly better out of the box experience.

Closes #54